@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/core.
  *
- * (c) 2012-2019 The MetaModels team.
+ * (c) 2012-2020 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -17,155 +17,127 @@
  * @author     Sven Baumann <baumann.sv@gmail.com>
  * @author     Richard Henkenjohann <richardhenkenjohann@googlemail.com>
  * @author     Ingolf Steinhardt <info@e-spin.de>
- * @copyright  2012-2019 The MetaModels team.
+ * @copyright  2012-2020 The MetaModels team.
  * @license    https://github.com/MetaModels/core/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
 
-$GLOBALS['TL_DCA']['tl_metamodel_dca_combine'] = array
-(
-    'config'     => array
-    (
+$GLOBALS['TL_DCA']['tl_metamodel_dca_combine'] = [
+    'config'     => [
         'dataContainer'    => 'General',
         'ptable'           => 'tl_metamodel',
         'switchToEdit'     => false,
         'enableVersioning' => false,
         'closed'           => false,
-        'sql'              => array
-        (
-            'keys' => array
-            (
+        'sql'              => [
+            'keys' => [
                 'id'       => 'primary',
                 'pid'      => 'index',
                 'fe_group' => 'index',
                 'be_group' => 'index'
-            ),
-        )
-    ),
-    'dca_config' => array
-    (
-        'data_provider' => array
-        (
-            'default' => array
-            (
+            ],
+        ]
+    ],
+    'dca_config' => [
+        'data_provider' => [
+            'default' => [
                 'class'        => 'ContaoCommunityAlliance\DcGeneral\Data\TableRowsAsRecordsDataProvider',
                 'source'       => 'tl_metamodel_dca_combine',
                 'group_column' => 'pid',
                 'sort_column'  => 'sorting'
-            )
-        ),
-    ),
-    'palettes'   => array
-    (
+            ]
+        ],
+    ],
+    'palettes'   => [
         'default' => 'rows'
-    ),
-    'fields'     => array
-    (
-        'id'       => array
-        (
+    ],
+    'fields'     => [
+        'id'       => [
             'sql' => 'int(10) unsigned NOT NULL auto_increment'
-        ),
-        'pid'      => array
-        (
+        ],
+        'pid'      => [
             'sql' => "int(10) unsigned NOT NULL default '0'"
-        ),
-        'sorting'  => array
-        (
+        ],
+        'sorting'  => [
             'sql' => "int(10) unsigned NOT NULL default '0'"
-        ),
-        'tstamp'   => array
-        (
+        ],
+        'tstamp'   => [
             'sql' => "int(10) unsigned NOT NULL default '0'"
-        ),
-        'rows'     => array
-        (
+        ],
+        'rows'     => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_dca_combine']['dca_combiner'],
             'exclude'   => true,
             'inputType' => 'multiColumnWizard',
-            'eval'      => array
-            (
+            'eval'      => [
                 'tl_class'     => 'dca_combine',
-                'columnFields' => array
-                (
-                    'id'       => array
-                    (
+                'columnFields' => [
+                    'id'       => [
                         'label'     => null,
                         'exclude'   => true,
                         'inputType' => 'justtext',
-                        'eval'      => array
-                        (
+                        'eval'      => [
                             'hideHead' => true,
                             'hideBody' => true,
-                        )
-                    ),
-                    'fe_group' => array
-                    (
+                        ]
+                    ],
+                    'fe_group' => [
                         'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_dca_combine']['fe_group'],
                         'exclude'   => true,
                         'inputType' => 'select',
-                        'eval'      => array
-                        (
+                        'eval'      => [
                             'includeBlankOption' => true,
-                            'style'              => 'width:115px',
+                            'style'              => 'width:100%',
                             'chosen'             => 'true'
-                        ),
-                    ),
-                    'be_group' => array
-                    (
+                        ],
+                    ],
+                    'be_group' => [
                         'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_dca_combine']['be_group'],
                         'exclude'   => true,
                         'inputType' => 'select',
-                        'eval'      => array
-                        (
+                        'eval'      => [
                             'includeBlankOption' => true,
-                            'style'              => 'width:115px',
+                            'style'              => 'width:100%',
                             'chosen'             => 'true'
-                        ),
-                    ),
-                    'dca_id'   => array
-                    (
-                        'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_dca_combine']['dca_id'],
-                        'exclude'   => true,
-                        'inputType' => 'select',
-                        'eval'      => array
-                        (
-                            'includeBlankOption' => true,
-                            'style'  => 'width:180px',
-                            'chosen' => 'true'
-                        ),
-                    ),
-                    'view_id'  => array
-                    (
+                        ],
+                    ],
+                    'view_id'  => [
                         'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_dca_combine']['view_id'],
                         'exclude'   => true,
                         'inputType' => 'select',
-                        'eval'      => array
-                        (
+                        'eval'      =>
+                            [
                             'includeBlankOption' => true,
-                            'style'  => 'width:180px',
-                            'chosen' => 'true'
-                        ),
-                    ),
-                ),
-            ),
-        ),
-        'fe_group' => array
-        (
+                            'style'              => 'width:100%',
+                            'chosen'             => 'true'
+                            ],
+                    ],
+                    'dca_id'   => [
+                        'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_dca_combine']['dca_id'],
+                        'exclude'   => true,
+                        'inputType' => 'select',
+                        'eval'      =>
+                            [
+                                'includeBlankOption' => true,
+                                'style'              => 'width:100%',
+                                'chosen'             => 'true'
+                            ],
+                    ],
+                ],
+            ],
+        ],
+        'fe_group' => [
             // keep signed as anonymous are -1
             'sql' => "int(10) NOT NULL default '0'"
-        ),
-        'be_group' => array
-        (
+        ],
+        'be_group' => [
             // keep signed as admins are -1
             'sql' => "int(10) NOT NULL default '0'"
-        ),
-        'dca_id'   => array
-        (
+        ],
+        'view_id'  => [
             'sql' => "int(10) unsigned NOT NULL default '0'"
-        ),
-        'view_id'  => array
-        (
+        ],
+        'dca_id'   => [
             'sql' => "int(10) unsigned NOT NULL default '0'"
-        )
-    )
-);
+        ]
+    ]
+];
