@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/core.
  *
- * (c) 2012-2019 The MetaModels team.
+ * (c) 2012-2020 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -21,192 +21,149 @@
  * @author     Cliff Parnitzky <github@cliff-parnitzky.de>
  * @author     Sven Baumann <baumann.sv@gmail.com>
  * @author     Ingolf Steinhardt <info@e-spin.de>
- * @copyright  2012-2019 The MetaModels team.
+ * @copyright  2012-2020 The MetaModels team.
  * @license    https://github.com/MetaModels/core/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
 
-$GLOBALS['TL_DCA']['tl_metamodel_dcasetting'] = array
-(
-    'config'                => array
-    (
+$GLOBALS['TL_DCA']['tl_metamodel_dcasetting'] = [
+    'config'                => [
         'dataContainer'    => 'General',
         'switchToEdit'     => true,
         'enableVersioning' => false,
-        'sql'              => array
-        (
-            'keys' => array
-            (
+        'sql'              => [
+            'keys' => [
                 'id'  => 'primary',
                 'pid' => 'index'
-            ),
-        ),
-    ),
-    'dca_config'            => array
-    (
-        'data_provider'  => array
-        (
-            'default'      => array
-            (
+            ],
+        ],
+    ],
+    'dca_config'            => [
+        'data_provider'  => [
+            'default'      => [
                 'source' => 'tl_metamodel_dcasetting'
-            ),
-            'parent'       => array
-            (
+            ],
+            'parent'       => [
                 'source' => 'tl_metamodel_dca'
-            ),
-            'root'         => array
-            (
+            ],
+            'root'         => [
                 'source' => 'tl_metamodel_dca'
-            ),
-            'tl_metamodel' => array
-            (
+            ],
+            'tl_metamodel' => [
                 'source' => 'tl_metamodel'
-            )
-        ),
-        'childCondition' => array
-        (
-            array
-            (
-                'from'   => 'tl_metamodel_dca',
-                'to'     => 'tl_metamodel_dcasetting',
-                'setOn'  => array
-                (
-                    array
-                    (
+            ]
+        ],
+        'childCondition' => [
+            [
+                'from'    => 'tl_metamodel_dca',
+                'to'      => 'tl_metamodel_dcasetting',
+                'setOn'   => [
+                    [
                         'to_field'   => 'pid',
                         'from_field' => 'id',
-                    ),
-                ),
-                'filter' => array
-                (
-                    array
-                    (
+                    ],
+                ],
+                'filter'  => [
+                    [
                         'local'     => 'pid',
                         'remote'    => 'id',
                         'operation' => '=',
-                    ),
-                ),
-                'inverse' => array
-                (
-                    array
-                    (
+                    ],
+                ],
+                'inverse' => [
+                    [
                         'local'     => 'pid',
                         'remote'    => 'id',
                         'operation' => '=',
-                    ),
-                )
-            ),
-            array
-            (
+                    ],
+                ]
+            ],
+            [
                 'from'    => 'tl_metamodel',
                 'to'      => 'tl_metamodel_dca',
-                'setOn'   => array
-                (
-                    array
-                    (
+                'setOn'   => [
+                    [
                         'to_field'   => 'pid',
                         'from_field' => 'id',
-                    ),
-                ),
-                'filter'  => array
-                (
-                    array
-                    (
+                    ],
+                ],
+                'filter'  => [
+                    [
                         'local'     => 'pid',
                         'remote'    => 'id',
                         'operation' => '=',
-                    ),
-                ),
-                'inverse' => array
-                (
-                    array
-                    (
+                    ],
+                ],
+                'inverse' => [
+                    [
                         'local'     => 'pid',
                         'remote'    => 'id',
                         'operation' => '=',
-                    ),
-                )
-            )
-        ),
-        'rootEntries'    => array
-        (
-            'tl_metamodel_dcasetting' => array
-            (
-                'setOn'  => array
-                (
-                    array
-                    (
+                    ],
+                ]
+            ]
+        ],
+        'rootEntries'    => [
+            'tl_metamodel_dcasetting' => [
+                'setOn'  => [
+                    [
                         'property' => 'pid',
                         'value'    => '0'
-                    ),
-                ),
-                'filter' => array
-                (
-                    array
-                    (
+                    ],
+                ],
+                'filter' => [
+                    [
                         'property'  => 'pid',
                         'operation' => '=',
                         'value'     => '0'
-                    )
-                )
-            )
-        ),
-        'child_list'     => array
-        (
-            'tl_metamodel_dcasetting' => array
-            (
-                'fields' => array
-                (
+                    ]
+                ]
+            ]
+        ],
+        'child_list'     => [
+            'tl_metamodel_dcasetting' => [
+                'fields' => [
                     'type',
                     'attr_id',
                     'urlparam',
                     'comment'
-                ),
+                ],
                 'format' => '%s %s',
-            ),
-        ),
-    ),
-    'list'                  => array
-    (
-        'sorting'           => array
-        (
+            ],
+        ],
+    ],
+    'list'                  => [
+        'sorting'           => [
             'mode'         => 4,
-            'fields'       => array('sorting'),
+            'fields'       => ['sorting'],
             'panelLayout'  => 'limit',
-            'headerFields' => array('name'),
-        ),
-        'global_operations' => array
-        (
-            'addall' => array
-            (
+            'headerFields' => ['name'],
+        ],
+        'global_operations' => [
+            'addall' => [
                 'label'      => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['addall'],
                 'class'      => 'header_add_all',
                 'attributes' => 'onclick="Backend.getScrollOffset();"'
-            ),
-            'all'    => array
-            (
+            ],
+            'all'    => [
                 'label'      => &$GLOBALS['TL_LANG']['MSC']['all'],
                 'href'       => 'act=select',
                 'class'      => 'header_edit_all',
                 'attributes' => 'onclick="Backend.getScrollOffset();"'
-            ),
-        ),
-        'operations'        => array
-        (
-            'edit'       => array
-            (
+            ],
+        ],
+        'operations'        => [
+            'edit'       => [
                 'label' => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['edit'],
                 'href'  => 'act=edit',
                 'icon'  => 'edit.svg'
-            ),
-            'cut'        => array
-            (
+            ],
+            'cut'        => [
                 'label' => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['cut'],
                 'href'  => 'act=paste&amp;mode=cut',
                 'icon'  => 'cut.svg'
-            ),
-            'delete'     => array
-            (
+            ],
+            'delete'     => [
                 'label'      => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['delete'],
                 'href'       => 'act=delete',
                 'icon'       => 'delete.svg',
@@ -214,85 +171,68 @@ $GLOBALS['TL_DCA']['tl_metamodel_dcasetting'] = array
                     'onclick="if (!confirm(\'%s\')) return false; Backend.getScrollOffset();"',
                     $GLOBALS['TL_LANG']['MSC']['deleteConfirm']
                 )
-            ),
-            'show'       => array
-            (
+            ],
+            'show'       => [
                 'label' => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['show'],
                 'href'  => 'act=show',
                 'icon'  => 'show.svg'
-            ),
-            'toggle'     => array
-            (
+            ],
+            'toggle'     => [
                 'label'          => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['toggle'],
                 'icon'           => 'visible.svg',
                 'toggleProperty' => 'published',
-            ),
-            'conditions' => array
-            (
+            ],
+            'conditions' => [
                 'label'   => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['conditions'],
                 'href'    => 'table=tl_metamodel_dcasetting_condition',
                 'icon'    => 'bundles/metamodelscore/images/icons/dca_condition.png',
                 'idparam' => 'pid'
-            ),
-        )
-    ),
-    'palettes'              => array
-    (
-        '__selector__' => array
-        (
+            ],
+        ]
+    ],
+    'palettes'              => [
+        '__selector__' => [
             'dcatype',
             'attr_id'
-        )
-    ),
-    'metapalettes'          => array
-    (
-        'default' => array
-        (
-            'title' => array
-            (
+        ]
+    ],
+    'metapalettes'          => [
+        'default' => [
+            'title' => [
                 'dcatype'
-            ),
-        ),
-    ),
-    'metasubselectpalettes' => array
-    (
-        'dcatype' => array
-        (
-            'attribute' => array
-            (
-                'title'     => array
-                (
+            ],
+        ],
+    ],
+    'metasubselectpalettes' => [
+        'dcatype' => [
+            'attribute' => [
+                'title'     => [
                     'attr_id'
-                ),
-                'functions' => array
-                (
+                ],
+                'functions' => [
                     'readonly'
-                ),
-                'advanced'  => array
-                ()
-            ),
-            'legend'    => array
-            (
-                'title' => array
-                (
+                ],
+                'advanced'  => []
+            ],
+            'legend'    => [
+                'title' => [
                     'legendhide',
                     'legendtitle',
-                )
-            )
-        ),
-        'attr_id' => array
-        (
+                ]
+            ]
+        ],
+        'attr_id' => [
             // Core legends:
             // * title
-            // * backend
-            // * config
+            // * functions
             // * advanced
             // Core fields:
             // * tl_class           css class to use in backend.
-            // * mandatory          mandatory
-            // * chosen
-            // * filterable         can be filtered (in backend)
-            // * searchable         can be searched (in backend)
+            // * mandatory          mandatory.
+            // * alwaysSave         always save.
+            // * filterable         can be filtered (in backend).
+            // * searchable         can be searched (in backend).
+            // * chosen             chosen for select.
             // * allowHtml          do not strip html content.
             // * preserveTags       do not encode html tags.
             // * decodeEntities     do decode HTML entities.
@@ -302,273 +242,234 @@ $GLOBALS['TL_DCA']['tl_metamodel_dcasetting'] = array
             // * trailingSlash      allow trailing slash, 2 => do nothing, 1 => add one on save, 0 => strip it on save.
             // * spaceToUnderscore  if true any whitespace character will be replaced by an underscore.
             // * includeBlankOption if true a blank option will be added to the options array.
-        )
-    ),
-    'fields'                => array
-    (
-        'id'                 => array
-        (
+            // * submitOnChange     submit on change value.
+            // * readonly           readonly.
+        ]
+    ],
+    'fields'                => [
+        'id'                 => [
             'sql' => 'int(10) unsigned NOT NULL auto_increment'
-        ),
-        'pid'                => array
-        (
+        ],
+        'pid'                => [
             'sql' => "int(10) unsigned NOT NULL default '0'"
-        ),
-        'sorting'            => array
-        (
+        ],
+        'sorting'            => [
             'sql' => "int(10) unsigned NOT NULL default '0'"
-        ),
-        'tstamp'             => array
-        (
+        ],
+        'tstamp'             => [
             'sql' => "int(10) unsigned NOT NULL default '0'"
-        ),
-        'published'          => array
-        (
-            'sql' => "char(1) NOT NULL default ''"
-        ),
-        'dcatype'            => array
-        (
+        ],
+        'published'          => [
+            'default' => 1,
+            'sql'     => "char(1) NOT NULL default ''"
+        ],
+        'dcatype'            => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['dcatype'],
             'exclude'   => true,
             'inputType' => 'select',
-            'options'   => array('attribute', 'legend'),
+            'options'   => ['attribute', 'legend'],
             'reference' => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['dcatypes'],
-            'eval'      => array
-            (
+            'eval'      => [
                 'tl_class'           => 'w50',
                 'includeBlankOption' => true,
                 'submitOnChange'     => true,
-            ),
+            ],
             'sql'       => "varchar(10) NOT NULL default ''"
-        ),
-        'attr_id'            => array
-        (
+        ],
+        'attr_id'            => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['attr_id'],
             'exclude'   => true,
             'inputType' => 'select',
-            'eval'      => array(
+            'eval'      => [
                 'tl_class'           => 'w50',
                 'doNotSaveEmpty'     => true,
                 'alwaysSave'         => true,
                 'includeBlankOption' => true,
                 'mandatory'          => true,
                 'submitOnChange'     => true,
-            ),
+            ],
             'sql'       => "int(10) unsigned NOT NULL default '0'"
-        ),
-        'tl_class'           => array
-        (
-            'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['tl_class'],
-            'exclude'   => true,
-            'inputType' => 'text',
-            'eval'      => array
-            (
-                'tl_class' => 'wizard',
-            ),
-            'sql'       => "varchar(64) NOT NULL default ''"
-        ),
-        'legendhide'         => array
-        (
+        ],
+        'tl_class'           => [
+            'label'       => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['tl_class'],
+            'exclude'     => true,
+            'inputType'   => 'text',
+            'eval'        => [
+                'tl_class'   => 'clr w50',
+                'helpwizard' => true,
+            ],
+            'explanation' => 'tl_class',
+            'sql'         => "varchar(64) NOT NULL default ''"
+        ],
+        'legendhide'         => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['legendhide'],
             'exclude'   => true,
             'inputType' => 'checkbox',
-            'eval'      => array
-            (
+            'eval'      => [
                 'tl_class' => 'w50 m12 cbx'
-            ),
+            ],
             'sql'       => "varchar(5) NOT NULL default ''"
-        ),
-        'legendtitle'        => array
-        (
+        ],
+        'legendtitle'        => [
             'label'   => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['legendtitle'],
             'exclude' => true,
-            'eval'    => array
-            (
+            'eval'    => [
                 'tl_class' => 'clr'
-            ),
+            ],
             'sql'     => 'text NULL'
-        ),
-        'mandatory'          => array
-        (
-            'label'                => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['mandatory'],
-            'exclude'              => true,
-            'inputType'            => 'checkbox',
-            'eval'                 => array
-            (
-                'tl_class' => 'w50',
-            ),
+        ],
+        'mandatory'          => [
+            'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['mandatory'],
+            'exclude'   => true,
+            'inputType' => 'checkbox',
+            'eval'      => [
+                'tl_class' => 'w50 cbx',
+            ],
             'sql'       => "char(1) NOT NULL default ''"
-        ),
-        'alwaysSave'         => array
-        (
+        ],
+        'alwaysSave'         => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['alwaysSave'],
             'exclude'   => true,
             'inputType' => 'checkbox',
-            'eval'      => array
-            (
+            'eval'      => [
                 'tl_class' => 'w50',
-            ),
+            ],
             'sql'       => "char(1) NOT NULL default ''"
-        ),
-        'filterable'         => array
-        (
+        ],
+        'filterable'         => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['filterable'],
             'exclude'   => true,
             'inputType' => 'checkbox',
-            'eval'      => array
-            (
-                'tl_class' => 'w50',
-            ),
+            'eval'      => [
+                'tl_class' => 'w50 cbx',
+            ],
             'sql'       => "char(1) NOT NULL default ''"
-        ),
-        'searchable'         => array
-        (
+        ],
+        'searchable'         => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['searchable'],
             'exclude'   => true,
             'inputType' => 'checkbox',
-            'eval'      => array
-            (
-                'tl_class' => 'w50',
-            ),
+            'eval'      => [
+                'tl_class' => 'w50 cbx',
+            ],
             'sql'       => "char(1) NOT NULL default ''"
-        ),
-        'chosen'             => array
-        (
+        ],
+        'chosen'             => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['chosen'],
             'exclude'   => true,
             'inputType' => 'checkbox',
-            'eval'      => array
-            (
-                'tl_class' => 'w50 m12'
-            ),
+            'eval'      => [
+                'tl_class' => 'w50 cbx'
+            ],
             'sql'       => "char(1) NOT NULL default ''"
-        ),
-        'allowHtml'          => array
-        (
+        ],
+        'allowHtml'          => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['allowHtml'],
             'exclude'   => true,
             'inputType' => 'checkbox',
-            'eval'      => array
-            (
-                'tl_class' => 'w50',
-            ),
+            'eval'      =>
+                [
+                    'tl_class' => 'w50 cbx',
+                ],
             'sql'       => "char(1) NOT NULL default ''"
-        ),
-        'preserveTags'       => array
-        (
+        ],
+        'preserveTags'       => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['preserveTags'],
             'exclude'   => true,
             'inputType' => 'checkbox',
-            'eval'      => array
-            (
-                'tl_class' => 'w50',
-            ),
+            'eval'      => [
+                'tl_class' => 'w50 cbx',
+            ],
             'sql'       => "char(1) NOT NULL default ''"
-        ),
-        'decodeEntities'     => array
-        (
+        ],
+        'decodeEntities'     => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['decodeEntities'],
             'exclude'   => true,
             'inputType' => 'checkbox',
-            'eval'      => array
-            (
-                'tl_class' => 'w50',
-            ),
+            'eval'      => [
+                'tl_class' => 'w50 cbx',
+            ],
             'sql'       => "char(1) NOT NULL default ''"
-        ),
-        'rte'                => array
-        (
+        ],
+        'rte'                => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['rte'],
             'exclude'   => true,
             'inputType' => 'select',
             'default'   => 'tinyMCE',
-            'eval'      => array
-            (
-                'tl_class'           => 'm12',
+            'eval'      => [
+                'tl_class'           => 'w50 m12',
                 'includeBlankOption' => true,
-            ),
+            ],
             'sql'       => "varchar(64) NOT NULL default ''"
-        ),
-        'rows'               => array
-        (
+        ],
+        'rows'               => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['rows'],
             'exclude'   => true,
             'inputType' => 'text',
-            'eval'      => array
-            (
-                'tl_class' => 'w50',
-                'rgxp'     => 'digit'
-            ),
+            'eval'      =>
+                [
+                    'tl_class' => 'w50',
+                    'rgxp'     => 'digit'
+                ],
             'sql'       => "int(10) NOT NULL default '0'"
-        ),
-        'cols'               => array
-        (
+        ],
+        'cols'               => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['cols'],
             'exclude'   => true,
             'inputType' => 'text',
-            'eval'      => array
-            (
+            'eval'      => [
                 'tl_class' => 'w50',
                 'rgxp'     => 'digit'
-            ),
+            ],
             'sql'       => "int(10) NOT NULL default '0'"
-        ),
-        'trailingSlash'      => array
-        (
+        ],
+        'trailingSlash'      => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['trailingSlash'],
             'exclude'   => true,
             'inputType' => 'select',
-            'options'   => array(0, 1, 2),
+            'options'   => [0, 1, 2],
             'default'   => 2,
             'reference' => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['trailingSlash_options'],
-            'eval'      => array
-            (
-                'tl_class' => 'w50 clr',
-            ),
+            'eval'      => [
+                'tl_class' => 'clr w50',
+            ],
             'sql'       => "char(1) NOT NULL default '2'"
-        ),
-        'spaceToUnderscore'  => array
-        (
+        ],
+        'spaceToUnderscore'  => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['spaceToUnderscore'],
             'exclude'   => true,
             'inputType' => 'checkbox',
-            'eval'      => array
-            (
-                'tl_class' => 'w50 m12',
-            ),
+            'eval'      => [
+                'tl_class' => 'w50 cbx m12',
+            ],
             'sql'       => "char(1) NOT NULL default ''"
-        ),
-        'includeBlankOption' => array
-        (
+        ],
+        'includeBlankOption' => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['includeBlankOption'],
             'exclude'   => true,
             'inputType' => 'checkbox',
-            'eval'      => array
-            (
-                'tl_class' => 'clr m12',
-            ),
+            'default'   => 1,
+            'eval'      => [
+                'tl_class' => 'clr cbx',
+            ],
             'sql'       => "char(1) NOT NULL default ''"
-        ),
-        'submitOnChange'     => array
-        (
+        ],
+        'submitOnChange'     => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['submitOnChange'],
             'exclude'   => true,
             'inputType' => 'checkbox',
-            'eval'      => array
-            (
-                'tl_class' => 'clr m12',
-            ),
+            'eval'      => [
+                'tl_class' => 'clr cbx',
+            ],
             'sql'       => "char(1) NOT NULL default ''"
-        ),
-        'readonly'           => array
-        (
+        ],
+        'readonly'           => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting']['readonly'],
             'inputType' => 'checkbox',
-            'eval'      => array
-            (
-                'tl_class' => 'w50',
-            ),
+            'eval'      => [
+                'tl_class' => 'w50 cbx',
+            ],
             'sql'       => "char(1) NOT NULL default ''"
-        )
-    )
-);
+        ]
+    ]
+];
