@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/core.
  *
- * (c) 2012-2019 The MetaModels team.
+ * (c) 2012-2020 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -17,7 +17,7 @@
  * @author     Ingolf Steinhardt <info@e-spin.de>
  * @author     Richard Henkenjohann <richardhenkenjohann@googlemail.com>
  * @author     Sven Baumann <baumann.sv@gmail.com>
- * @copyright  2012-2019 The MetaModels team.
+ * @copyright  2012-2020 The MetaModels team.
  * @license    https://github.com/MetaModels/core/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
@@ -26,181 +26,143 @@
  * Table tl_metamodel_dcasetting_condition
  */
 
-$GLOBALS['TL_DCA']['tl_metamodel_dcasetting_condition'] = array
-(
-    'config'                => array
-    (
+$GLOBALS['TL_DCA']['tl_metamodel_dcasetting_condition'] = [
+    'config'                => [
         'dataContainer'    => 'General',
+        'label'            => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting_condition']['list_label'],
         'switchToEdit'     => false,
         'enableVersioning' => false,
-        'sql'              => array
-        (
-            'keys' => array
-            (
+        'sql'              => [
+            'keys' => [
                 'id'  => 'primary',
                 'pid' => 'index'
-            ),
-        ),
-    ),
-    'dca_config'            => array
-    (
-        'data_provider'  => array
-        (
-            'root'   => array
-            (
+            ],
+        ],
+    ],
+    'dca_config'            => [
+        'data_provider'  => [
+            'root'   => [
                 'source' => 'tl_metamodel_dcasetting_condition'
-            ),
-            'parent' => array
-            (
+            ],
+            'parent' => [
                 'source' => 'tl_metamodel_dcasetting',
-            ),
-        ),
-        'childCondition' => array
-        (
-            array
-            (
+            ],
+        ],
+        'childCondition' => [
+            [
                 'from'   => 'tl_metamodel_dcasetting',
                 'to'     => 'tl_metamodel_dcasetting_condition',
-                'setOn'  => array
-                (
-                    array
-                    (
+                'setOn'  => [
+                    [
                         'to_field'   => 'settingId',
                         'from_field' => 'id',
-                    )
-                ),
-                'filter' => array
-                (
-                    array
-                    (
+                    ]
+                ],
+                'filter' => [
+                    [
                         'local'     => 'settingId',
                         'remote'    => 'id',
                         'operation' => '=',
-                    ),
-                )
-            ),
-            array(
+                    ],
+                ]
+            ],
+            [
                 'from'   => 'tl_metamodel_dcasetting_condition',
                 'to'     => 'tl_metamodel_dcasetting_condition',
-                'setOn'  => array
-                (
-                    array
-                    (
+                'setOn'  => [
+                    [
                         'to_field'   => 'pid',
                         'from_field' => 'id',
-                    ),
-                    array
-                    (
+                    ],
+                    [
                         'to_field'   => 'settingId',
                         'from_field' => 'settingId',
-                    ),
-                ),
-                'filter' => array
-                (
-                    array
-                    (
+                    ],
+                ],
+                'filter' => [
+                    [
                         'local'     => 'pid',
                         'remote'    => 'id',
                         'operation' => '=',
-                    ),
-                )
-            )
-        ),
-        'rootEntries'    => array
-        (
-            'tl_metamodel_dcasetting_condition' => array
-            (
-                'setOn'  => array
-                (
-                    array
-                    (
+                    ],
+                ]
+            ]
+        ],
+        'rootEntries'    => [
+            'tl_metamodel_dcasetting_condition' => [
+                'setOn'  => [
+                    [
                         'property' => 'pid',
                         'value'    => '0'
-                    ),
-                ),
-                'filter' => array
-                (
-                    array
-                    (
+                    ],
+                ],
+                'filter' => [
+                    [
                         'property'  => 'pid',
                         'operation' => '=',
                         'value'     => '0'
-                    )
-                )
-            )
-        ),
-        'child_list'     => array
-        (
-            'tl_metamodel_dcasetting_condition' => array
-            (
-                'fields' => array
-                (
+                    ]
+                ]
+            ]
+        ],
+        'child_list'     => [
+            'tl_metamodel_dcasetting_condition' => [
+                'fields' => [
                     'condition',
                     'attr_id',
                     'comment',
                     'enabled'
-                ),
+                ],
                 'format' => '%s %s',
-            ),
-        )
-    ),
-    'list'                  => array
-    (
-        'sorting'           => array
-        (
-            'mode'         => 5,
-            'fields'       => array('sorting'),
-            'headerFields' => array
-            (
-                'type',
-                'attr_id'
-            ),
-            'flag'         => 1,
-            'icon'         => 'bundles/metamodelscore/images/icons/filter_and.png',
-        ),
-        'label'             => array
-        (
-            'fields' => array
-            (
+            ],
+        ]
+    ],
+    'list'                  => [
+        'sorting'           =>
+            [
+                'mode'         => 5,
+                'fields'       => ['sorting'],
+                'headerFields' => [
+                    'type',
+                    'attr_id'
+                ],
+                'flag'         => 1,
+                'icon'         => 'bundles/metamodelscore/images/icons/filter_and.png',
+            ],
+        'label'             => [
+            'fields' => [
                 'type',
                 'attr_id',
                 'comment'
-            ),
+            ],
             'format' => '%s %s %s',
-        ),
-        'global_operations' => array
-        (
-            'all' => array
-            (
+        ],
+        'global_operations' => [
+            'all' => [
                 'label'      => &$GLOBALS['TL_LANG']['MSC']['all'],
                 'href'       => 'act=select',
                 'class'      => 'header_edit_all',
                 'attributes' => 'onclick="Backend.getScrollOffset();"'
-            )
-        ),
-        'operations'        => array
-        (
-            'edit'   => array
-            (
+            ]
+        ],
+        'operations'        => [
+            'edit'   => [
                 'label' => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting_condition']['edit'],
                 'href'  => 'act=edit',
                 'icon'  => 'edit.svg'
-            ),
-            'copy'   => array
-            (
+            ],
+            'copy'   => [
                 'label' => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting_condition']['copy'],
                 'href'  => 'act=copy',
                 'icon'  => 'copy.svg'
-            ),
-            'cut'    => array
-            (
+            ],
+            'cut'    => [
                 'label'      => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting_condition']['cut'],
                 'href'       => 'act=paste&amp;mode=cut',
                 'icon'       => 'cut.svg',
                 'attributes' => 'onclick="Backend.getScrollOffset()"',
-            ),
-            'delete' => array
-            (
+            ],
+            'delete' => [
                 'label'      => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting_condition']['delete'],
                 'href'       => 'act=delete',
                 'icon'       => 'delete.svg',
@@ -208,104 +170,78 @@ $GLOBALS['TL_DCA']['tl_metamodel_dcasetting_condition'] = array
                     'onclick="if (!confirm(\'%s\')) return false; Backend.getScrollOffset();"',
                     $GLOBALS['TL_LANG']['MSC']['deleteConfirm']
                 )
-            ),
-            'show'   => array
-            (
+            ],
+            'show'   => [
                 'label' => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting_condition']['show'],
                 'href'  => 'act=show',
                 'icon'  => 'show.svg'
-            ),
-            'toggle' => array
-            (
+            ],
+            'toggle' => [
                 'label'          => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting_condition']['toggle'],
                 'icon'           => 'visible.svg',
                 'toggleProperty' => 'enabled',
-            )
-        )
-    ),
-    'palettes'              => array
-    (
-        '__selector__' => array
-        (
+            ]
+        ]
+    ],
+    'palettes'              => [
+        '__selector__' => [
             'type'
-        )
-    ),
-    'metapalettes'          => array
-    (
-        'default'                                           => array
-        (
-            'basic' => array
-            (
+        ]
+    ],
+    'metapalettes'          => [
+        'default'                                           => [
+            'basic' => [
                 'type',
                 'enabled',
                 'comment'
-            ),
-        ),
-        '_attribute_ extends default'                       => array
-        (
-            '+config' => array
-            (
+            ],
+        ],
+        '_attribute_ extends default'                       => [
+            '+config' => [
                 'attr_id'
-            )
-        ),
-        'conditionor extends default'                       => array
-        (),
-        'conditionand extends default'                      => array
-        (),
-        'conditionpropertyvalueis extends _attribute_'      => array
-        (
-            '+config' => array
-            (
+            ]
+        ],
+        'conditionor extends default'                       => [],
+        'conditionand extends default'                      => [],
+        'conditionpropertyvalueis extends _attribute_'      => [
+            '+config' => [
                 'value'
-            )
-        ),
-        'conditionpropertycontainanyof extends _attribute_' => array
-        (
-            '+config' => array
-            (
+            ]
+        ],
+        'conditionpropertycontainanyof extends _attribute_' => [
+            '+config' => [
                 'value'
-            )
-        ),
-        'conditionpropertyvisible extends _attribute_'      => array
-        (),
-    ),
-    'metasubselectpalettes' => array
-    (
-        'attr_id' => array
-        ()
-    ),
-    'fields'                => array
-    (
-        'id'        => array
-        (
+            ]
+        ],
+        'conditionpropertyvisible extends _attribute_'      => [],
+    ],
+    'metasubselectpalettes' => [
+        'attr_id' => []
+    ],
+    'fields'                => [
+        'id'        => [
             'sql' => 'int(10) unsigned NOT NULL auto_increment'
-        ),
-        'pid'       => array
-        (
+        ],
+        'pid'       => [
             'sql' => "int(10) unsigned NOT NULL default '0'"
-        ),
-        'sorting'   => array
-        (
+        ],
+        'sorting'   => [
             'sql' => "int(10) unsigned NOT NULL default '0'"
-        ),
-        'tstamp'    => array
-        (
+        ],
+        'tstamp'    => [
             'sql' => "int(10) unsigned NOT NULL default '0'"
-        ),
-        'settingId' => array
-        (
+        ],
+        'settingId' => [
             // Keep this empty but keep it here!
             // needed for act=copy in DC_Table, as otherwise the fid value will not be copied.
             'label' => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting_condition']['fid'],
             'sql'   => "int(10) unsigned NOT NULL default '0'"
-        ),
-        'type'      => array
-        (
+        ],
+        'type'      => [
             'label'       => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting_condition']['type'],
             'exclude'     => true,
             'inputType'   => 'select',
-            'eval'        => array
-            (
+            'eval'        => [
                 'doNotSaveEmpty'     => true,
                 'alwaysSave'         => true,
                 'submitOnChange'     => true,
@@ -314,40 +250,35 @@ $GLOBALS['TL_DCA']['tl_metamodel_dcasetting_condition'] = array
                 'tl_class'           => 'w50',
                 'chosen'             => true,
                 'helpwizard'         => true
-            ),
+            ],
             'explanation' => 'dcasetting_condition',
             'sql'         => "varchar(255) NOT NULL default ''"
-        ),
-        'enabled'   => array
-        (
+        ],
+        'enabled'   => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting_condition']['enabled'],
             'exclude'   => true,
             'inputType' => 'checkbox',
-            'eval'      => array
-            (
+            'default'   => 1,
+            'eval'      => [
                 'alwaysSave' => true,
-                'tl_class'   => 'w50 m12',
-            ),
+                'tl_class'   => 'w50 m12 cbx',
+            ],
             'sql'       => "char(1) NOT NULL default ''"
-        ),
-        'comment'   => array
-        (
+        ],
+        'comment'   => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting_condition']['comment'],
             'exclude'   => true,
             'inputType' => 'text',
-            'eval'      => array
-            (
+            'eval'      => [
                 'tl_class' => 'clr long'
-            ),
+            ],
             'sql'       => "varchar(255) NOT NULL default ''"
-        ),
-        'attr_id'   => array
-        (
+        ],
+        'attr_id'   => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting_condition']['attr_id'],
             'exclude'   => true,
             'inputType' => 'select',
-            'eval'      => array
-            (
+            'eval'      => [
                 'doNotSaveEmpty'     => true,
                 'alwaysSave'         => true,
                 'submitOnChange'     => true,
@@ -355,22 +286,20 @@ $GLOBALS['TL_DCA']['tl_metamodel_dcasetting_condition'] = array
                 'mandatory'          => true,
                 'tl_class'           => 'w50',
                 'chosen'             => true
-            ),
+            ],
             'sql'       => "int(10) unsigned NOT NULL default '0'"
-        ),
-        'value'     => array
-        (
+        ],
+        'value'     => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_dcasetting_condition']['value'],
             'exclude'   => true,
             'inputType' => 'select',
-            'eval'      => array
-            (
+            'eval'      => [
                 'alwaysSave'         => true,
                 'includeBlankOption' => true,
                 'tl_class'           => 'w50',
                 'chosen'             => true
-            ),
+            ],
             'sql'       => 'blob NULL'
-        ),
-    )
-);
+        ],
+    ]
+];
