@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/core.
  *
- * (c) 2012-2019 The MetaModels team.
+ * (c) 2012-2020 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -20,7 +20,7 @@
  * @author     Ingolf Steinhardt <info@e-spin.de>
  * @author     Richard Henkenjohann <richardhenkenjohann@googlemail.com>
  * @author     Sven Baumann <baumann.sv@gmail.com>
- * @copyright  2012-2019 The MetaModels team.
+ * @copyright  2012-2020 The MetaModels team.
  * @license    https://github.com/MetaModels/core/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
@@ -29,178 +29,142 @@
  * Table tl_metamodel_attribute
  */
 
-$GLOBALS['TL_DCA']['tl_metamodel_filtersetting'] = array
-(
-    'config'                => array
-    (
+$GLOBALS['TL_DCA']['tl_metamodel_filtersetting'] = [
+    'config'                => [
         'dataContainer'    => 'General',
+        'label'            => &$GLOBALS['TL_LANG']['tl_metamodel_filtersetting']['list_label'],
         'switchToEdit'     => false,
         'enableVersioning' => false,
-        'sql'              => array
-        (
-            'keys' => array
-            (
+        'sql'              => [
+            'keys' => [
                 'id'  => 'primary',
                 'pid' => 'index'
-            ),
-        )
-    ),
-    'dca_config'            => array
-    (
-        'data_provider'  => array
-        (
-            'root'   => array
-            (
+            ],
+        ]
+    ],
+    'dca_config'            => [
+        'data_provider'  => [
+            'root'   => [
                 'source' => 'tl_metamodel_filtersetting'
-            ),
-            'parent' => array
-            (
+            ],
+            'parent' => [
                 'source' => 'tl_metamodel_filter',
-            ),
-        ),
-        'childCondition' => array
-        (
-            array(
+            ],
+        ],
+        'childCondition' => [
+            [
                 'from'   => 'tl_metamodel_filter',
                 'to'     => 'tl_metamodel_filtersetting',
-                'setOn'  => array
-                (
-                    array(
+                'setOn'  => [
+                    [
                         'to_field'   => 'fid',
                         'from_field' => 'id',
-                    )
-                ),
-                'filter' => array
-                (
-                    array
-                    (
+                    ]
+                ],
+                'filter' => [
+                    [
                         'local'     => 'fid',
                         'remote'    => 'id',
                         'operation' => '=',
-                    ),
-                )
-            ),
-            array(
+                    ],
+                ]
+            ],
+            [
                 'from'   => 'tl_metamodel_filtersetting',
                 'to'     => 'tl_metamodel_filtersetting',
-                'setOn'  => array
-                (
-                    array
-                    (
+                'setOn'  => [
+                    [
                         'to_field'   => 'pid',
                         'from_field' => 'id',
-                    ),
-                    array
-                    (
+                    ],
+                    [
                         'to_field'   => 'fid',
                         'from_field' => 'fid',
-                    ),
-                ),
-                'filter' => array
-                (
-                    array
-                    (
+                    ],
+                ],
+                'filter' => [
+                    [
                         'local'     => 'pid',
                         'remote'    => 'id',
                         'operation' => '=',
-                    ),
-                )
-            )
-        ),
-        'rootEntries'    => array
-        (
-            'tl_metamodel_filtersetting' => array
-            (
-                'setOn'  => array
-                (
-                    array
-                    (
+                    ],
+                ]
+            ]
+        ],
+        'rootEntries'    => [
+            'tl_metamodel_filtersetting' => [
+                'setOn'  => [
+                    [
                         'property' => 'pid',
                         'value'    => '0'
-                    ),
-                ),
-                'filter' => array
-                (
-                    array
-                    (
+                    ],
+                ],
+                'filter' => [
+                    [
                         'property'  => 'pid',
                         'operation' => '=',
                         'value'     => '0'
-                    )
-                )
-            )
-        ),
-        'child_list'     => array
-        (
-            'tl_metamodel_filtersetting' => array
-            (
-                'fields' => array
-                (
+                    ]
+                ]
+            ]
+        ],
+        'child_list'     => [
+            'tl_metamodel_filtersetting' => [
+                'fields' => [
                     'type',
                     'attr_id',
                     'urlparam',
                     'comment',
                     'enabled'
-                ),
+                ],
                 'format' => '%s %s',
-            ),
-        )
-    ),
-    'list'                  => array
-    (
-        'sorting'           => array
-        (
+            ],
+        ]
+    ],
+    'list'                  => [
+        'sorting'           => [
             'mode'         => 5,
-            'fields'       => array('sorting'),
-            'headerFields' => array('type', 'attr_id'),
+            'fields'       => ['sorting'],
+            'headerFields' => ['type', 'attr_id'],
             'flag'         => 1,
             'icon'         => 'bundles/metamodelscore/images/icons/filter_and.png',
-        ),
-        'label'             => array
-        (
-            'fields' => array
-            (
+        ],
+        'label'             => [
+            'fields' => [
                 'fid',
                 'type',
                 'attr_id',
                 'urlparam',
                 'comment'
-            ),
+            ],
             'format' => '%s',
-        ),
-        'global_operations' => array
-        (
-            'all' => array
-            (
+        ],
+        'global_operations' => [
+            'all' => [
                 'label'      => &$GLOBALS['TL_LANG']['MSC']['all'],
                 'href'       => 'act=select',
                 'class'      => 'header_edit_all',
                 'attributes' => 'onclick="Backend.getScrollOffset();"'
-            )
-        ),
-        'operations'        => array
-        (
-            'edit'   => array
-            (
+            ]
+        ],
+        'operations'        => [
+            'edit'   => [
                 'label' => &$GLOBALS['TL_LANG']['tl_metamodel_filtersetting']['edit'],
                 'href'  => 'act=edit',
                 'icon'  => 'edit.svg'
-            ),
-            'copy'   => array
-            (
+            ],
+            'copy'   => [
                 'label' => &$GLOBALS['TL_LANG']['tl_metamodel_filtersetting']['copy'],
                 'href'  => 'act=copy',
                 'icon'  => 'copy.svg'
-            ),
-            'cut'    => array
-            (
+            ],
+            'cut'    => [
                 'label'      => &$GLOBALS['TL_LANG']['tl_metamodel_filtersetting']['cut'],
                 'href'       => 'act=paste&amp;mode=cut',
                 'icon'       => 'cut.svg',
                 'attributes' => 'onclick="Backend.getScrollOffset()"',
-            ),
-            'delete' => array
-            (
+            ],
+            'delete' => [
                 'label'      => &$GLOBALS['TL_LANG']['tl_metamodel_filtersetting']['delete'],
                 'href'       => 'act=delete',
                 'icon'       => 'delete.svg',
@@ -208,63 +172,49 @@ $GLOBALS['TL_DCA']['tl_metamodel_filtersetting'] = array
                     'onclick="if (!confirm(\'%s\')) return false; Backend.getScrollOffset();"',
                     $GLOBALS['TL_LANG']['MSC']['deleteConfirm']
                 )
-            ),
-            'show'   => array
-            (
+            ],
+            'show'   => [
                 'label' => &$GLOBALS['TL_LANG']['tl_metamodel_filtersetting']['show'],
                 'href'  => 'act=show',
                 'icon'  => 'show.svg'
-            ),
-            'toggle' => array
-            (
+            ],
+            'toggle' => [
                 'label'          => &$GLOBALS['TL_LANG']['tl_metamodel_filtersetting']['toggle'],
                 'icon'           => 'visible.svg',
                 'toggleProperty' => 'enabled',
-            )
-        )
-    ),
-    'palettes'              => array
-    (
-        '__selector__' => array
-        (
+            ]
+        ]
+    ],
+    'palettes'              => [
+        '__selector__' => [
             'type'
-        )
-    ),
-    'metapalettes'          => array
-    (
-        'default'                                      => array
-        (
-            'title' => array(
+        ]
+    ],
+    'metapalettes'          => [
+        'default'                          => [
+            'title' => [
                 'type',
                 'enabled',
                 'comment'
-            ),
-        ),
-        '_attribute_ extends default'                  => array
-        (
-            'config' => array
-            (
+            ],
+        ],
+        '_attribute_ extends default'      => [
+            'config' => [
                 'attr_id'
-            )
-        ),
-        'conditionor extends default'                  => array
-        (
-            'config' => array
-            (
+            ]
+        ],
+        'conditionor extends default'      => [
+            'config' => [
                 'stop_after_match'
-            )
-        ),
-        'idlist extends default'                       => array
-        (
-            '+config' => array
-            (
+            ]
+        ],
+        'idlist extends default'           => [
+            '+config' => [
                 'items'
-            ),
-        ),
-        'simplelookup extends _attribute_'             => array
-        (
-            '+fefilter' => array
-            (
+            ],
+        ],
+        'simplelookup extends _attribute_' => [
+            '+fefilter' => [
                 'urlparam',
                 'predef_param',
                 'fe_widget',
@@ -276,64 +226,49 @@ $GLOBALS['TL_DCA']['tl_metamodel_filtersetting'] = array
                 'onlyused',
                 'onlypossible',
                 'skipfilteroptions'
-            ),
-        ),
-        'customsql extends default'                    => array
-        (
-            '+config' => array
-            (
+            ],
+        ],
+        'customsql extends default'        => [
+            '+config' => [
                 'customsql'
-            ),
-        )
-    ),
-    'metasubselectpalettes' => array
-    (
-        'attr_id' => array
-        ()
-    ),
-    'simplelookup_palettes' => array
-    (
-        '_translated_' => array
-        (
-            'config' => array
-            (
+            ],
+        ]
+    ],
+    'metasubselectpalettes' => [
+        'attr_id' => []
+    ],
+    'simplelookup_palettes' => [
+        '_translated_' => [
+            'config' => [
                 'all_langs'
-            )
-        )
-    ),
-    'fields'                => array
-    (
-        'id'                => array
-        (
+            ]
+        ]
+    ],
+    'fields'                => [
+        'id'                => [
             'sql' => 'int(10) unsigned NOT NULL auto_increment'
-        ),
-        'pid'               => array
-        (
+        ],
+        'pid'               => [
             'sql' => "int(10) unsigned NOT NULL default '0'"
-        ),
-        'sorting'           => array
-        (
+        ],
+        'sorting'           => [
             'sorting' => true,
             'sql'     => "int(10) unsigned NOT NULL default '0'"
-        ),
-        'tstamp'            => array
-        (
+        ],
+        'tstamp'            => [
             'sql' => "int(10) unsigned NOT NULL default '0'"
-        ),
-        'fid'               => array
-        (
+        ],
+        'fid'               => [
             // Keep this empty but keep it here!
             // needed for act=copy in DC_Table, as otherwise the fid value will not be copied.
             'label' => &$GLOBALS['TL_LANG']['tl_metamodel_filtersetting']['fid'],
             'sql'   => "int(10) unsigned NOT NULL default '0'"
-        ),
-        'type'              => array
-        (
+        ],
+        'type'              => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_filtersetting']['type'],
             'exclude'   => true,
             'inputType' => 'select',
-            'eval'      => array
-            (
+            'eval'      => [
                 'doNotSaveEmpty'     => true,
                 'alwaysSave'         => true,
                 'submitOnChange'     => true,
@@ -341,36 +276,32 @@ $GLOBALS['TL_DCA']['tl_metamodel_filtersetting'] = array
                 'mandatory'          => true,
                 'tl_class'           => 'w50',
                 'chosen'             => true
-            ),
+            ],
             'sql'       => "varchar(64) NOT NULL default ''"
-        ),
-        'enabled'           => array
-        (
+        ],
+        'enabled'           => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_filtersetting']['enabled'],
             'exclude'   => true,
             'inputType' => 'checkbox',
-            'eval'      => array
-            (
+            'default'   => 1,
+            'eval'      => [
                 'alwaysSave' => true,
-                'tl_class'   => 'w50 m12',
-            ),
+                'tl_class'   => 'w50 m12 cbx',
+            ],
             'sql'       => "char(1) NOT NULL default ''"
-        ),
-        'comment'           => array
-        (
+        ],
+        'comment'           => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_filtersetting']['comment'],
             'exclude'   => true,
             'inputType' => 'text',
-            'eval'      => array('tl_class' => 'clr long'),
+            'eval'      => ['tl_class' => 'clr long'],
             'sql'       => "varchar(255) NOT NULL default ''"
-        ),
-        'attr_id'           => array
-        (
+        ],
+        'attr_id'           => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_filtersetting']['attr_id'],
             'exclude'   => true,
             'inputType' => 'select',
-            'eval'      => array
-            (
+            'eval'      => [
                 'doNotSaveEmpty'     => true,
                 'alwaysSave'         => true,
                 'submitOnChange'     => true,
@@ -378,194 +309,164 @@ $GLOBALS['TL_DCA']['tl_metamodel_filtersetting'] = array
                 'mandatory'          => true,
                 'tl_class'           => 'w50',
                 'chosen'             => true
-            ),
+            ],
             'sql'       => "int(10) unsigned NOT NULL default '0'"
-        ),
-        'all_langs'         => array
-        (
+        ],
+        'all_langs'         => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_filtersetting']['all_langs'],
             'exclude'   => true,
             'inputType' => 'checkbox',
-            'eval'      => array
-            (
+            'eval'      => [
                 'alwaysSave' => true,
                 'tl_class'   => 'w50 m12 cbx',
-            ),
+            ],
             'sql'       => "char(1) NOT NULL default ''"
-        ),
-        'items'             => array
-        (
+        ],
+        'items'             => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_filtersetting']['items'],
             'exclude'   => true,
             'inputType' => 'textarea',
-            'eval'      => array
-            (
+            'eval'      => [
                 'doNotSaveEmpty' => true,
                 'alwaysSave'     => true,
                 'mandatory'      => true,
-            ),
+            ],
             'sql'       => 'text NULL'
-        ),
-        'urlparam'          => array
-        (
+        ],
+        'urlparam'          => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_filtersetting']['urlparam'],
             'exclude'   => true,
             'inputType' => 'text',
-            'eval'      => array
-            (
+            'eval'      => [
                 'tl_class' => 'w50',
-            ),
+            ],
             'sql'       => "varchar(255) NOT NULL default ''"
-        ),
-        'predef_param'      => array
-        (
+        ],
+        'predef_param'      => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_filtersetting']['predef_param'],
             'exclude'   => true,
             'inputType' => 'checkbox',
-            'eval'      => array
-            (
+            'eval'      => [
                 'alwaysSave' => true,
-                'tl_class'   => 'w50 m12',
-            ),
+                'tl_class'   => 'w50 m12 cbx',
+            ],
             'sql'       => "char(1) NOT NULL default ''"
-        ),
-        'fe_widget'      => array
-        (
+        ],
+        'fe_widget'         => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_filtersetting']['fe_widget'],
             'exclude'   => true,
             'inputType' => 'checkbox',
-            'eval'      => array
-            (
+            'eval'      => [
                 'alwaysSave' => true,
-                'tl_class'   => 'w50 m12',
-            ),
+                'tl_class'   => 'w50 m12 cbx',
+            ],
             'sql'       => "char(1) NOT NULL default ''"
-        ),
-        'customsql'         => array
-        (
+        ],
+        'customsql'         => [
             'label'       => &$GLOBALS['TL_LANG']['tl_metamodel_filtersetting']['customsql'],
             'exclude'     => true,
             'inputType'   => 'textarea',
-            'eval'        => array
-            (
+            'eval'        => [
                 'allowHtml'      => true,
                 'preserveTags'   => true,
                 'decodeEntities' => true,
                 'rte'            => 'ace|sql',
                 'class'          => 'monospace',
                 'helpwizard'     => true,
-            ),
+            ],
             'explanation' => 'customsql',
             'sql'         => 'text NULL'
-        ),
-        'allow_empty'       => array
-        (
+        ],
+        'allow_empty'       => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_filtersetting']['allow_empty'],
             'exclude'   => true,
             'inputType' => 'checkbox',
-            'eval'      => array
-            (
+            'eval'      => [
                 'alwaysSave' => true,
-                'tl_class'   => 'w50 m12',
-            ),
+                'tl_class'   => 'w50 m12 cbx',
+            ],
             'sql'       => "char(1) NOT NULL default ''"
-        ),
-        'stop_after_match'  => array
-        (
+        ],
+        'stop_after_match'  => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_filtersetting']['stop_after_match'],
             'exclude'   => true,
             'inputType' => 'checkbox',
-            'eval'      => array
-            (
+            'eval'      => [
                 'alwaysSave' => true,
-                'tl_class'   => 'w50 m12',
-            ),
+                'tl_class'   => 'w50 cbx',
+            ],
             'sql'       => "char(1) NOT NULL default ''"
-        ),
-        'label'             => array
-        (
+        ],
+        'label'             => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_filtersetting']['label'],
             'exclude'   => true,
             'inputType' => 'text',
-            'eval'      => array
-            (
+            'eval'      => [
                 'tl_class' => 'clr w50',
-            ),
+            ],
             'sql'       => 'blob NULL'
-        ),
-        'template'          => array
-        (
+        ],
+        'template'          => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_filtersetting']['template'],
             'default'   => 'mm_filteritem_default',
             'exclude'   => true,
             'inputType' => 'select',
-            'eval'      => array
-            (
+            'eval'      => [
                 'tl_class' => 'w50',
                 'chosen'   => true
-            ),
+            ],
             'sql'       => "varchar(64) NOT NULL default ''"
-        ),
-        'blankoption'       => array
-        (
+        ],
+        'blankoption'       => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_filtersetting']['blankoption'],
             'exclude'   => true,
             'default'   => true,
             'inputType' => 'checkbox',
-            'eval'      => array
-            (
-                'tl_class' => 'w50 clr',
-            ),
+            'eval'      => [
+                'tl_class' => 'clr w50 m12 cbx',
+            ],
             'sql'       => "char(1) NOT NULL default '1'"
-        ),
-        'onlyused'          => array
-        (
+        ],
+        'onlyused'          => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_filtersetting']['onlyused'],
             'exclude'   => true,
             'default'   => true,
             'inputType' => 'checkbox',
-            'eval'      => array
-            (
-                'tl_class'       => 'w50',
+            'eval'      => [
+                'tl_class'       => 'w50 m12 cbx',
                 'submitOnChange' => true,
-            ),
+            ],
             'sql'       => "char(1) NOT NULL default '0'"
-        ),
-        'onlypossible'      => array
-        (
+        ],
+        'onlypossible'      => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_filtersetting']['onlypossible'],
             'exclude'   => true,
             'default'   => true,
             'inputType' => 'checkbox',
-            'eval'      => array
-            (
-                'tl_class' => 'w50',
-            ),
+            'eval'      => [
+                'tl_class' => 'w50 m12 cbx',
+            ],
             'sql'       => "char(1) NOT NULL default '0'"
-        ),
-        'skipfilteroptions' => array
-        (
+        ],
+        'skipfilteroptions' => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_filtersetting']['skipfilteroptions'],
             'exclude'   => true,
             'default'   => false,
             'inputType' => 'checkbox',
-            'eval'      => array
-            (
-                'tl_class' => 'w50',
-            ),
+            'eval'      => [
+                'tl_class' => 'w50 cbx',
+            ],
             'sql'       => "char(1) NOT NULL default '0'"
-        ),
-        'defaultid'         => array
-        (
+        ],
+        'defaultid'         => [
             'label'     => &$GLOBALS['TL_LANG']['tl_metamodel_filtersetting']['defaultid'],
             'exclude'   => true,
             'inputType' => 'select',
-            'eval'      => array
-            (
-                'tl_class'           => 'w50 clr',
+            'eval'      => [
+                'tl_class'           => 'clr w50',
                 'includeBlankOption' => true
-            ),
+            ],
             'sql'       => "varchar(255) NOT NULL default ''"
-        )
-    )
-);
+        ]
+    ]
+];
