@@ -3,7 +3,7 @@
 /**
  * This file is part of MetaModels/core.
  *
- * (c) 2012-2018 The MetaModels team.
+ * (c) 2012-2020 The MetaModels team.
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -15,7 +15,7 @@
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
  * @author     Sven Baumann <baumann.sv@gmail.com>
  * @author     Ingolf Steinhardt <info@e-spin.de>
- * @copyright  2012-2018 The MetaModels team.
+ * @copyright  2012-2020 The MetaModels team.
  * @license    https://github.com/MetaModels/core/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
  */
@@ -23,6 +23,7 @@
 /**
  * Fields
  */
+$GLOBALS['TL_LANG']['tl_metamodel_filtersetting']['list_label']                = 'Filter settings';
 $GLOBALS['TL_LANG']['tl_metamodel_filtersetting']['fid'][0]                    = 'Parent collection';
 $GLOBALS['TL_LANG']['tl_metamodel_filtersetting']['fid'][1]                    =
     'The collection of filter settings, this setting belongs to.';
