@@ -18,6 +18,7 @@
  * @author     David Maack <david.maack@arcor.de>
  * @author     Oliver Hoff <oliver@hofff.com>
  * @author     Stefan Heimes <stefan_heimes@hotmail.com>
+ * @author     Richard Henkenjohann <richardhenkenjohann@googlemail.com>
  * @copyright  2012-2019 The MetaModels team.
  * @license    https://github.com/MetaModels/core/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
@@ -109,15 +110,6 @@ class HybridList extends MetaModelHybrid
         }
 
         $objItemRenderer
-            ->setServiceContainerFallback(
-                function () {
-                    return $this->getServiceContainer();
-                }
-            )
-            ->setFactory(System::getContainer()->get('metamodels.factory'))
-            ->setFilterFactory(System::getContainer()->get('metamodels.filter_setting_factory'))
-            ->setRenderSettingFactory(System::getContainer()->get('metamodels.render_setting_factory'))
-            ->setEventDispatcher(System::getContainer()->get('event_dispatcher'))
             ->setMetaModel($this->metamodel, $this->metamodel_rendersettings)
             ->setLimit($this->metamodel_use_limit, $this->metamodel_offset, $this->metamodel_limit)
             ->setPageBreak($this->perPage)
