@@ -14,6 +14,7 @@
  * @author     Christian Schiffler <c.schiffler@cyberspectrum.de>
  * @author     Sven Baumann <baumann.sv@gmail.com>
  * @author     David Molineus <david.molineus@netzmacht.de>
+ * @author     Richard Henkenjohann <richardhenkenjohann@googlemail.com>
  * @copyright  2012-2019 The MetaModels team.
  * @license    https://github.com/MetaModels/core/blob/master/LICENSE LGPL-3.0-or-later
  * @filesource
@@ -21,6 +22,8 @@
 
 namespace MetaModels;
 
+use Contao\BackendUser;
+use Contao\FrontendUser;
 use Doctrine\Common\Cache\Cache;
 use MetaModels\Attribute\IAttributeFactory;
 use MetaModels\Filter\Setting\IFilterSettingFactory;
@@ -396,7 +399,7 @@ class MetaModelsServiceContainer implements IMetaModelsServiceContainer
                     $this->services['metamodels-view-combinations'] =
                         new \MetaModels\FrontendIntegration\ViewCombinations(
                             $this,
-                            $GLOBALS['container']['user'],
+                            FrontendUser::getInstance(),
                             \System::getContainer()->get('database_connection')
                         );
                     break;
@@ -404,7 +407,7 @@ class MetaModelsServiceContainer implements IMetaModelsServiceContainer
                     $this->services['metamodels-view-combinations'] =
                         new \MetaModels\BackendIntegration\ViewCombinations(
                             $this,
-                            $GLOBALS['container']['user'],
+                            BackendUser::getInstance(),
                             \System::getContainer()->get('database_connection')
                         );
                     break;
